src/requirements.txt
django
django-ninja
channels[daphne]
python-dotenv
httpx
django-cors-headers
channels_redis